+++ hahomematic/const.py
@@ -9,7 +9,7 @@ from enum import Enum, IntEnum, StrEnum
 import re
 from typing import Any, Final, Required, TypedDict
 
-VERSION: Final = "2024.11.2"
+VERSION: Final = "2024.11.3"
 
 DEFAULT_CONNECTION_CHECKER_INTERVAL: Final = 15  # check if connection is available via rpc ping
 DEFAULT_CUSTOM_ID: Final = "custom_id"
@@ -88,7 +88,7 @@ SYSVAR_STATE_PATH_ROOT: Final = "sysvar/status"
 PROGRAM_SET_PATH_ROOT: Final = "program/set"
 PROGRAM_STATE_PATH_ROOT: Final = "program/status"
 
-MAX_CACHE_AGE: Final = 60
+MAX_CACHE_AGE: Final = 15
 
 NO_CACHE_ENTRY: Final = "NO_CACHE_ENTRY"