+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.11.2 (2024-11-15)
 
+- Add get_data_point_path to central
 - Ignore unknown interfaces
 - Remove clients for not available interfaces